apps/api/management/commands/init_tokens.py
from django.core.management.base import BaseCommand, CommandError
from django.contrib.auth.models import User
from rest_framework.authtoken.models import Token

class Command(BaseCommand):
    help = 'Initializes authentication tokens for all users in the database, if they do not not exist'

    def handle(self, *args, **options):
        for user in User.objects.all():
            Token.objects.get_or_create(user=user)

        self.stdout.write('Successfully initialized tokens.')

