requirements.txt
Django==1.10.2
django-randomslugfield==0.3.0
MySQL-python==1.2.5
Pillow==3.4.2
djangorestframework==3.5.3
django-filter==1.0.1
Markdown==2.6.7